+++ GenumKit/Stencil/Contexts.swift
@@ -34,7 +34,7 @@ private extension String {
     - `alpha`: `String` — hex value of the alpha component
 */
 extension ColorsFileParser {
-  public func stencilContext(enumName: String = "ColorName") -> Context {
+  public func stencilContext(enumName: String = "ColorName") -> [String: Any] {
     let colorMap = colors.map({ (color: (name: String, value: UInt32)) -> [String:String] in
       let name = color.name.trimmingCharacters(in: CharacterSet.whitespaces)
       let hex = "00000000" + String(color.value, radix: 16)
@@ -51,7 +51,7 @@ extension ColorsFileParser {
         "alpha": comps[3]
       ]
     }).sorted { $0["name"] ?? "" < $1["name"] ?? "" }
-    return Context(dictionary: ["enumName": enumName, "colors": colorMap], namespace: genumNamespace())
+    return ["enumName": enumName, "colors": colorMap]
   }
 }
 
@@ -61,18 +61,20 @@ extension ColorsFileParser {
  - `images`: `Array<String>` — list of image names
 */
 extension AssetsCatalogParser {
-  public func stencilContext(enumName: String = "Asset") -> Context {
-    let images = justValues(entries: entries)
-    let imagesStructured = structure(entries: entries)
+  public func stencilContext(enumName: String = "Asset") -> [String: Any] {
+    let images = catalogs.flatMap { justValues(entries: $1) }.sorted(by: <)
+    let structured = catalogs.keys.sorted(by: <).map { name -> [String: Any] in
+      return [
+        "name": name,
+        "assets": structure(entries: catalogs[name] ?? [])
+      ]
+    }
 
-    return Context(
-      dictionary: [
-        "enumName": enumName,
-        "images": images,
-        "structuredImages": imagesStructured
-      ],
-      namespace: genumNamespace()
-    )
+    return [
+      "enumName": enumName,
+      "images": images,
+      "catalogs": structured
+    ]
   }
 
   private func justValues(entries: [Entry]) -> [String] {
@@ -80,7 +82,7 @@ extension AssetsCatalogParser {
 
     for entry in entries {
       switch entry {
-      case let .namespace(name: _, items: items):
+      case let .group(name: _, items: items):
         result += justValues(entries: items)
       case let .image(name: _, value: value):
         result += [value]
@@ -90,21 +92,14 @@ extension AssetsCatalogParser {
     return result
   }
 
-  private func structure(entries: [Entry], currentLevel: Int = 0, maxLevel: Int = 5) -> [[String: Any]] {
+  private func structure(entries: [Entry]) -> [[String: Any]] {
     return entries.map { entry in
       switch entry {
-      case let .namespace(name: name, items: items):
-        if currentLevel + 1 >= maxLevel {
-          return [
-            "name": name,
-            "items": flatten(entries: items)
-          ]
-        } else {
-          return [
-            "name": name,
-            "items": structure(entries: items, currentLevel: currentLevel + 1, maxLevel: maxLevel)
-          ]
-        }
+      case let .group(name: name, items: items):
+        return [
+          "name": name,
+          "items": structure(entries: items)
+        ]
       case let .image(name: name, value: value):
         return [
           "name": name,
@@ -113,36 +108,13 @@ extension AssetsCatalogParser {
       }
     }
   }
-
-  private func flatten(entries: [Entry]) -> [[String: Any]] {
-    var result = [[String: Any]]()
-
-    for entry in entries {
-      switch entry {
-      case let .namespace(name: name, items: items):
-        result += flatten(entries: items).map { item in
-          return [
-            "name": "\(name)/\(item["name"]!)",
-            "value": item["value"]!
-          ]
-        }
-      case let .image(name: name, value: value):
-        result += [[
-          "name": name,
-          "value": value
-        ]]
-      }
-    }
-
-    return result
-  }
 }
 
 /* MARK: - Stencil Context for Storyboards
 
  - `sceneEnumName`: `String`
  - `segueEnumName`: `String`
- - `extraImports`: `Array` of `String`
+ - `modules`: `Array` of `String`
  - `storyboards`: `Array` of:
     - `name`: `String`
     - `initialScene`: `Dictionary` (absent if not specified)
@@ -163,8 +135,7 @@ extension AssetsCatalogParser {
 */
 extension StoryboardParser {
   public func stencilContext(sceneEnumName: String = "StoryboardScene",
-                                           segueEnumName: String = "StoryboardSegue",
-                                           extraImports: [String] = []) -> Context {
+                                           segueEnumName: String = "StoryboardSegue") -> [String: Any] {
     let storyboards = Set(storyboardsScenes.keys).union(storyboardsSegues.keys).sorted(by: <)
     let storyboardsMap = storyboards.map { (storyboardName: String) -> [String:Any] in
       var sbMap: [String:Any] = ["name": storyboardName]
@@ -172,7 +143,7 @@ extension StoryboardParser {
       if let initialScene = initialScenes[storyboardName] {
         let initial: [String:Any]
         if let customClass = initialScene.customClass {
-          initial = ["customClass": customClass]
+          initial = ["customClass": customClass, "customModule": initialScene.customModule ?? ""]
         } else {
           initial = [
             "baseType": uppercaseFirst(initialScene.tag),
@@ -187,36 +158,36 @@ extension StoryboardParser {
           .sorted(by: {$0.storyboardID < $1.storyboardID})
           .map { (scene: Scene) -> [String:Any] in
             if let customClass = scene.customClass {
-                return ["identifier": scene.storyboardID, "customClass": customClass]
+              return [
+                "identifier": scene.storyboardID,
+                "customClass": customClass,
+                "customModule": scene.customModule ?? ""
+              ]
             } else if scene.tag == "viewController" {
-                return [
-                    "identifier": scene.storyboardID,
-                    "baseType": uppercaseFirst(scene.tag),
-                    "isBaseViewController": scene.tag == "viewController"
-                ]
+              return [
+                "identifier": scene.storyboardID,
+                "baseType": uppercaseFirst(scene.tag),
+                "isBaseViewController": scene.tag == "viewController"
+              ]
             }
             return ["identifier": scene.storyboardID, "baseType": uppercaseFirst(scene.tag)]
         }
       }
       // All Segues
       if let segues = storyboardsSegues[storyboardName] {
-        sbMap["segues"] = segues
-          .sorted(by: {$0.segueID < $1.segueID})
-          .map { (segue: Segue) -> [String:String] in
-            ["identifier": segue.segueID, "customClass": segue.customClass ?? ""]
-        }
+        sbMap["segues"] = segues.sorted(by: {$0.identifier < $1.identifier})
       }
       return sbMap
     }
-    return Context(
-      dictionary: [
-        "sceneEnumName": sceneEnumName,
-        "segueEnumName": segueEnumName,
-        "extraImports": extraImports,
-        "storyboards": storyboardsMap
-      ],
-      namespace: genumNamespace()
-    )
+    return [
+      "sceneEnumName": sceneEnumName,
+      "segueEnumName": segueEnumName,
+      "storyboards": storyboardsMap,
+      "modules": modules.sorted(),
+
+      // NOTE: This is a deprecated variable
+      "extraImports": modules.sorted()
+    ]
   }
 }
 
@@ -238,7 +209,7 @@ extension StoryboardParser {
  - `structuredStrings`: `Dictionary` - contains strings structured by keys separated by '.' syntax
 */
 extension StringsFileParser {
-  public func stencilContext(enumName: String = "L10n", tableName: String = "Localizable") -> Context {
+  public func stencilContext(enumName: String = "L10n", tableName: String = "Localizable") -> [String: Any] {
 
     let entryToStringMapper = { (entry: Entry, keyPath: [String]) -> [String: Any] in
       var keyStructure = entry.keyStructure
@@ -271,20 +242,15 @@ extension StringsFileParser {
         .map { entryToStringMapper($0, []) }
     let structuredStrings = structure(
         entries: entries,
-        usingMapper: entryToStringMapper,
-        currentLevel: 0,
-        maxLevel: 5
+        usingMapper: entryToStringMapper
     )
 
-    return Context(dictionary:
-      [
-        "enumName": enumName,
-        "tableName": tableName,
-        "strings": strings,
-        "structuredStrings": structuredStrings
-      ],
-      namespace: genumNamespace()
-    )
+    return [
+      "enumName": enumName,
+      "tableName": tableName,
+      "strings": strings,
+      "structuredStrings": structuredStrings
+    ]
   }
 
   private func normalize(_ string: String) -> String {
@@ -296,9 +262,7 @@ extension StringsFileParser {
   private func structure(
     entries: [Entry],
     atKeyPath keyPath: [String] = [],
-    usingMapper mapper: @escaping Mapper,
-    currentLevel: Int,
-    maxLevel: Int) -> [String: Any] {
+    usingMapper mapper: @escaping Mapper) -> [String: Any] {
 
     var structuredStrings: [String: Any] = [:]
 
@@ -334,50 +298,15 @@ extension StringsFileParser {
       let entriesInKeyPath = entries.filter {
         Array($0.keyStructure.map(normalize).prefix(nextLevelKeyPath.count)) == nextLevelKeyPath.map(normalize)
       }
-      if currentLevel >= maxLevel {
-        subenums.append(
-            flattenedStrings(fromEnteries: entries,
-                             atKeyPath: nextLevelKeyPath,
-                             usingMapper: mapper,
-                             level: currentLevel+1
-            )
-        )
-      } else {
-        subenums.append(
-            structure(entries: entriesInKeyPath,
-                      atKeyPath: nextLevelKeyPath,
-                      usingMapper: mapper,
-                      currentLevel: currentLevel+1,
-                      maxLevel: maxLevel)
-        )
-      }
+      subenums.append(
+          structure(entries: entriesInKeyPath,
+                    atKeyPath: nextLevelKeyPath,
+                    usingMapper: mapper)
+      )
     }
 
     if !subenums.isEmpty {
       structuredStrings["subenums"] = subenums
-    }
-
-    return structuredStrings
-  }
-
-  private func flattenedStrings(
-    fromEnteries entries: [Entry],
-    atKeyPath keyPath: [String],
-    usingMapper mapper: @escaping Mapper,
-    level: Int) -> [String: Any] {
-
-    var structuredStrings: [String: Any] = [:]
-
-    let strings = entries
-      .filter { $0.keyStructure.count >= keyPath.count+1 }
-      .map { mapper($0, keyPath) }
-
-    if !strings.isEmpty {
-      structuredStrings["strings"] = strings
-    }
-
-    if let lastKeyPathComponent = keyPath.last {
-      structuredStrings["name"] = lastKeyPathComponent
     }
 
     return structuredStrings
@@ -394,7 +323,7 @@ extension StringsFileParser {
 */
 
 extension FontsFileParser {
-  public func stencilContext(enumName: String = "FontFamily") -> Context {
+  public func stencilContext(enumName: String = "FontFamily") -> [String: Any] {
     // turn into array of dictionaries
     let families = entries.map { (name: String, family: Set<Font>) -> [String: Any] in
       let fonts = family.map { (font: Font) -> [String: String] in
@@ -411,6 +340,6 @@ extension FontsFileParser {
       ]
     }.sorted { $0["name"] as? String ?? "" < $1["name"] as? String ?? "" }
 
-    return Context(dictionary: ["enumName": enumName, "families": families], namespace: genumNamespace())
+    return ["enumName": enumName, "families": families]
   }
 }