+++ UnitTests/TestSuites/ColorsTests.swift
@@ -15,10 +15,10 @@ class ColorsTextFileTests: XCTestCase {
   func testEmpty() {
     let parser = ColorsTextFileParser()
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-Empty.swift.out")
+    let expected = Fixtures.string(for: "Colors-Empty.swift.out")
     XCTDiffStrings(result, expected)
   }
 
@@ -32,10 +32,10 @@ class ColorsTextFileTests: XCTestCase {
       XCTFail("Failed with unexpected error \(error)")
     }
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-List-Default.swift.out")
+    let expected = Fixtures.string(for: "Colors-List-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
@@ -49,50 +49,50 @@ class ColorsTextFileTests: XCTestCase {
       XCTFail("Failed with unexpected error \(error)")
     }
 
-    let template = GenumTemplate(templateString: fixtureString("colors-rawValue.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-rawValue.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-List-RawValue.swift.out")
+    let expected = Fixtures.string(for: "Colors-List-RawValue.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDefaults() {
     let parser = ColorsTextFileParser()
-    try! parser.parseFile(at: fixture("colors.txt"))
+    try! parser.parseFile(at: Fixtures.path(for: "colors.txt"))
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-Txt-File-Default.swift.out")
+    let expected = Fixtures.string(for: "Colors-Txt-File-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileSwift3() {
     let parser = ColorsTextFileParser()
-    try! parser.parseFile(at: fixture("colors.txt"))
+    try! parser.parseFile(at: Fixtures.path(for: "colors.txt"))
 
-    let template = GenumTemplate(templateString: fixtureString("colors-swift3.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-swift3.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-Txt-File-Swift3.swift.out")
+    let expected = Fixtures.string(for: "Colors-Txt-File-Swift3.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithCustomName() {
     let parser = ColorsTextFileParser()
-    try! parser.parseFile(at: fixture("colors.txt"))
+    try! parser.parseFile(at: Fixtures.path(for: "colors.txt"))
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext(enumName: "XCTColors"))
 
-    let expected = fixtureString("Colors-Txt-File-CustomName.swift.out")
+    let expected = Fixtures.string(for: "Colors-Txt-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithBadSyntax() {
     let parser = ColorsTextFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad-syntax.txt"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad-syntax.txt"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
     } catch ColorsParserError.invalidFile {
       // That's the expected exception we want to happen
@@ -104,7 +104,7 @@ class ColorsTextFileTests: XCTestCase {
   func testFileWithBadValue() {
     let parser = ColorsTextFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad-value.txt"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad-value.txt"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad value")
     } catch ColorsParserError.invalidHexColor(string: "thisIsn'tAColor", key: "ArticleTitle"?) {
       // That's the expected exception we want to happen
@@ -122,40 +122,39 @@ class ColorsCLRFileTests: XCTestCase {
   func testEmpty() {
     let parser = ColorsCLRFileParser()
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-Empty.swift.out")
+    let expected = Fixtures.string(for: "Colors-Empty.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDefaults() {
     let parser = ColorsCLRFileParser()
-    try! parser.parseFile(at: fixture("colors.clr"))
+    try! parser.parseFile(at: Fixtures.path(for: "colors.clr"))
 
-
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-File-Default.swift.out")
+    let expected = Fixtures.string(for: "Colors-File-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithCustomName() {
     let parser = ColorsCLRFileParser()
-    try! parser.parseFile(at: fixture("colors.clr"))
+    try! parser.parseFile(at: Fixtures.path(for: "colors.clr"))
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext(enumName: "XCTColors"))
 
-    let expected = fixtureString("Colors-File-CustomName.swift.out")
+    let expected = Fixtures.string(for: "Colors-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithBadFile() {
     let parser = ColorsCLRFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad.clr"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad.clr"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad file")
     } catch ColorsParserError.invalidFile {
       // That's the expected exception we want to happen
@@ -171,47 +170,47 @@ class ColorsXMLFileTests: XCTestCase {
   func testEmpty() {
     let parser = ColorsXMLFileParser()
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-Empty.swift.out")
+    let expected = Fixtures.string(for: "Colors-Empty.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDefaults() {
     let parser = ColorsXMLFileParser()
     do {
-      try parser.parseFile(at: fixture("colors.xml"))
+      try parser.parseFile(at: Fixtures.path(for: "colors.xml"))
     } catch {
       XCTFail("Exception while parsing file: \(error)")
     }
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-File-Default.swift.out")
+    let expected = Fixtures.string(for: "Colors-File-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithCustomName() {
     let parser = ColorsXMLFileParser()
     do {
-      try parser.parseFile(at: fixture("colors.xml"))
+      try parser.parseFile(at: Fixtures.path(for: "colors.xml"))
     } catch {
       XCTFail("Exception while parsing file: \(error)")
     }
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext(enumName: "XCTColors"))
 
-    let expected = fixtureString("Colors-File-CustomName.swift.out")
+    let expected = Fixtures.string(for: "Colors-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithBadSyntax() {
     let parser = ColorsXMLFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad-syntax.xml"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad-syntax.xml"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
     } catch ColorsParserError.invalidFile {
       // That's the expected exception we want to happen
@@ -223,7 +222,7 @@ class ColorsXMLFileTests: XCTestCase {
   func testFileWithBadValue() {
     let parser = ColorsXMLFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad-value.xml"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad-value.xml"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad value")
     } catch ColorsParserError.invalidHexColor(string: "this isn't a color", key: "ArticleTitle"?) {
       // That's the expected exception we want to happen
@@ -239,47 +238,47 @@ class ColorsJSONFileTests: XCTestCase {
   func testEmpty() {
     let parser = ColorsJSONFileParser()
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-Empty.swift.out")
+    let expected = Fixtures.string(for: "Colors-Empty.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDefaults() {
     let parser = ColorsJSONFileParser()
     do {
-      try parser.parseFile(at: fixture("colors.json"))
+      try parser.parseFile(at: Fixtures.path(for: "colors.json"))
     } catch {
       XCTFail("Exception while parsing file: \(error)")
     }
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Colors-File-Default.swift.out")
+    let expected = Fixtures.string(for: "Colors-File-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithCustomName() {
     let parser = ColorsJSONFileParser()
     do {
-      try parser.parseFile(at: fixture("colors.json"))
+      try parser.parseFile(at: Fixtures.path(for: "colors.json"))
     } catch {
       XCTFail("Exception while parsing file: \(error)")
     }
 
-    let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "colors-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext(enumName: "XCTColors"))
 
-    let expected = fixtureString("Colors-File-CustomName.swift.out")
+    let expected = Fixtures.string(for: "Colors-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithBadSyntax() {
     let parser = ColorsJSONFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad-syntax.json"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad-syntax.json"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
     } catch ColorsParserError.invalidFile {
       // That's the expected exception we want to happen
@@ -291,7 +290,7 @@ class ColorsJSONFileTests: XCTestCase {
   func testFileWithBadValue() {
     let parser = ColorsJSONFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad-value.json"))
+      try parser.parseFile(at: Fixtures.path(for: "colors-bad-value.json"))
       XCTFail("Code did parse file successfully while it was expected to fail for bad value")
     } catch ColorsParserError.invalidHexColor(string: "this isn't a color", key: "ArticleTitle"?) {
       // That's the expected exception we want to happen