+++ UnitTests/TestSuites/StringsTests.swift
@@ -18,10 +18,10 @@ class StringsTests: XCTestCase {
   func testEmpty() {
     let parser = StringsFileParser()
 
-    let template = GenumTemplate(templateString: fixtureString("strings-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-Empty.swift.out")
+    let expected = Fixtures.string(for: "Strings-Empty.swift.out")
     XCTDiffStrings(result, expected)
   }
 
@@ -30,131 +30,131 @@ class StringsTests: XCTestCase {
     parser.addEntry(StringsFileParser.Entry(key: "Title", translation: "My awesome title"))
     parser.addEntry(StringsFileParser.Entry(key: "Greetings", translation: "Hello, my name is %@ and I'm %d", types: .Object, .Int))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-Entries-Default.swift.out")
+    let expected = Fixtures.string(for: "Strings-Entries-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDefaults() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-Default.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testMultiline() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("LocMultiline.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "LocMultiline.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-Multiline.swift.out")
+    let expected = Fixtures.string(for: "Strings-Multiline.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testUTF8FileWithDefaults() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("LocUTF8.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "LocUTF8.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-UTF8-Default.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-UTF8-Default.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithCustomName() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-default.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-default.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext(enumName: "XCTLoc"))
 
-    let expected = fixtureString("Strings-File-CustomName.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithSwift3() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-swift3.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-swift3.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-Swift3.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-Swift3.swift.out")
     XCTDiffStrings(result, expected)
   }
   
   func testFileWithoutCommentsAndSwift3() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
     
-    let template = GenumTemplate(templateString: fixtureString("strings-no-comments-swift3.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-no-comments-swift3.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
     
-    let expected = fixtureString("Strings-File-NoComments-Swift3.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-NoComments-Swift3.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithStructured() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-structured.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-structured.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-Structured.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-Structured.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithStructuredOnly() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("LocStructuredOnly.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "LocStructuredOnly.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-structured.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-structured.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-Structured-Only.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-Structured-Only.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDotSyntax() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-dot-syntax.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-dot-syntax.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-Dot-Syntax.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-Dot-Syntax.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithDotSyntaxSwift3() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
 
-    let template = GenumTemplate(templateString: fixtureString("strings-dot-syntax-swift3.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-dot-syntax-swift3.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
 
-    let expected = fixtureString("Strings-File-Dot-Syntax-Swift3.swift.out")
+    let expected = Fixtures.string(for: "Strings-File-Dot-Syntax-Swift3.swift.out")
     XCTDiffStrings(result, expected)
   }
 
   func testFileWithGenstringsTemplate() {
     let parser = StringsFileParser()
-    try! parser.parseFile(at: fixture("Localizable.strings"))
+    try! parser.parseFile(at: Fixtures.path(for: "Localizable.strings"))
     
-    let template = GenumTemplate(templateString: fixtureString("strings-genstrings.stencil"))
+    let template = GenumTemplate(templateString: Fixtures.string(for: "strings-genstrings.stencil"), environment: genumEnvironment())
     let result = try! template.render(parser.stencilContext())
     
-    let expected = fixtureString("Strings-Localizable-Genstrings.swift.out")
+    let expected = Fixtures.string(for: "Strings-Localizable-Genstrings.swift.out")
     XCTDiffStrings(result, expected)
   }