+++ swiftgen-cli/fonts.swift
@@ -13,15 +13,19 @@ let fontsCommand = command(
   outputOption,
   templateOption(prefix: "fonts"), templatePathOption,
   Option<String>("enumName", "FontFamily", flag: "e", description: "The name of the enum to generate"),
-  Argument<Path>("DIR", description: "Directory to parse.", validator: dirExists)
-  ) { output, templateName, templatePath, enumName, path in
+  VariadicArgument<Path>("DIR", description: "Directory to parse.", validator: dirsExist)
+  ) { output, templateName, templatePath, enumName, paths in
     let parser = FontsFileParser()
     do {
-      parser.parseFile(at: path)
+      for path in paths {
+        parser.parseFile(at: path)
+      }
+
       let templateRealPath = try findTemplate(
         prefix: "fonts", templateShortName: templateName, templateFullPath: templatePath
       )
-      let template = try GenumTemplate(path: templateRealPath)
+
+      let template = try GenumTemplate(templateString: templateRealPath.read(), environment: genumEnvironment())
       let context = parser.stencilContext(enumName: enumName)
       let rendered = try template.render(context)
       output.write(content: rendered, onlyIfChanged: true)