+++ swiftgen-cli/storyboards.swift
@@ -15,19 +15,22 @@ let storyboardsCommand = command(
     description: "The name of the enum to generate for Scenes"),
   Option<String>("segueEnumName", "StoryboardSegue", flag: "g",
     description: "The name of the enum to generate for Segues"),
+  // Note: import option is deprecated.
   VariadicOption<String>("import", [],
-    description: "Additional imports to be added to the generated file"),
-  Argument<Path>("PATH",
+    description: "Additional imports to be added to the generated file (DEPRECATED)"),
+  VariadicArgument<Path>("PATH",
     description: "Directory to scan for .storyboard files. Can also be a path to a single .storyboard",
-    validator: pathExists)
-) { output, templateName, templatePath, sceneEnumName, segueEnumName, extraImports, path in
+    validator: pathsExist)
+) { output, templateName, templatePath, sceneEnumName, segueEnumName, extraImports, paths in
   let parser = StoryboardParser()
 
   do {
-    if path.extension == "storyboard" {
-      try parser.addStoryboard(at: path)
-    } else {
-      try parser.parseDirectory(at: path)
+    for path in paths {
+      if path.extension == "storyboard" {
+        try parser.addStoryboard(at: path)
+      } else {
+        try parser.parseDirectory(at: path)
+      }
     }
 
     let templateRealPath = try findTemplate(
@@ -35,9 +38,9 @@ let storyboardsCommand = command(
       templateShortName: templateName,
       templateFullPath: templatePath
     )
-    let template = try GenumTemplate(path: templateRealPath)
+    let template = try GenumTemplate(templateString: templateRealPath.read(), environment: genumEnvironment())
     let context = parser.stencilContext(
-      sceneEnumName: sceneEnumName, segueEnumName: segueEnumName, extraImports: extraImports
+      sceneEnumName: sceneEnumName, segueEnumName: segueEnumName
     )
     let rendered = try template.render(context)
     output.write(content: rendered, onlyIfChanged: true)